Move DataTable to components/index/cards/charts
