+++ components/index/cards/ConfirmedCasesAttributesCard.vue
@@ -22,7 +22,8 @@
 import Data from '@/data/data.json'
 import formatGraph from '@/utils/formatGraph'
 import formatTable from '@/utils/formatTable'
-import DataTable from '@/components/DataTable.vue'
+
+import DataTable from '@/components/index/cards/charts/DataTable.vue'
 
 export default {
   components: {
